testing/requirements.txt
pytest
flake8
autopep8
azure-identity
azure-keyvault-secrets
azure-storage-file-datalake
azure-synapse
azure-synapse-artifacts >0.19.0
azure-mgmt-resource
pyodbc
pandas
pyarrow